Move under Platform Team ownership
